requirements.txt
Flask
Flask-SQLAlchemy
pillow
pytesseract
opencv-python
numpy